Create styles.css
This is a basic CSS style. For now keeping it here temporarily to get an idea how the index page would look. Expecting Md Akram Hossain to replace it with the one he working an I helped slightly. This current one is basically modified version of a previous testing css file.