initial commit for creating the quiz page
This html file contains the basic structure and main interface of the quiz page. As I am not done with the app.js server-side work yet, also have not pushed that yet, and the CSS file is not completely ready yet, this file should only show the skeleton version. To make it actually dynamic, and to make the questions appear, I am working on server-side part too, which I will push next. Also, on this file, have added a script with a relative part of 2 sound file, which I am trying to make working, but need to wait until I finish the app.js part and also choose and upload actual sound tracks.