styles.css testing version update 2.1
This is a temporary basic CSS style for testing purpose only. For now keeping it here temporarily to get an idea how the index.html and quiz.html, the 2 frint end html file I pushed already, would look. This is actually an modified version of the previous testing CSS, this modified one covers basic design for index and quiz page only. Expecting Md Akram Hossain to replace it with the one he working an I helped slightly, when he is done giving the final touch .